refactor(mcp): split lib/api/mcp.py into a namespace package
The monolithic 754-line lib/api/mcp.py mixed FastMCP setup, auth
helpers, URL builders, JSON serialization, and 11 tool definitions.
Split it into focused modules under lib/api/mcp/ so each file fits the
project's 300-line target and tools can be located by domain.

Layout:
  instance.py     FastMCP instance + mcp_auth
  server.py       Imports tool modules to register decorators; exports mcp_app
  helpers.py      resolve_user, URL builders, require_api_key, mint_tus_bearer_token
  serialization.py  get_project_details_json, build_gate_required_payload
  tools/          One module per tool domain (projects, workflows, files,
                  uploads, revisions, workflow_types)

Also drops underscore prefixes from helpers/serialization functions that
are now imported across modules — they're no longer module-private.

Updated lib/api/main.py and the two MCP unit-test files (imports +
mock.patch paths repointed to where each symbol is actually used).
