+++ tests/unit/test_mcp_tools.py
@@ -13,23 +13,31 @@ import lib.api.mcp_auth as _mcp_auth_mod
 _orig = _mcp_auth_mod.create_mcp_auth
 _mcp_auth_mod.create_mcp_auth = lambda: None
 
-from lib.api.mcp import (  # noqa: E402
-    _build_project_url,
-    _build_settings_url,
-    _build_tus_url,
-    _require_api_key,
-    _resolve_user,
-    create_project,
-    create_revision,
-    export_project_docx,
-    get_project,
-    get_tus_upload_credentials,
-    list_project_files,
-    list_projects,
-    list_revisions,
-    remove_reference_file,
-    run_workflow,
+from lib.api.mcp.helpers import (  # noqa: E402
+    build_project_url,
+    build_settings_url,
+    build_tus_url,
+    require_api_key,
+    resolve_user,
 )
+from lib.api.mcp.tools.files import (  # noqa: E402
+    export_project_docx,
+    list_project_files,
+    remove_reference_file,
+)
+from lib.api.mcp.tools.projects import (  # noqa: E402
+    create_project,
+    get_project,
+    list_projects,
+)
+from lib.api.mcp.tools.revisions import (  # noqa: E402
+    create_revision,
+    list_revisions,
+)
+from lib.api.mcp.tools.uploads import (  # noqa: E402
+    get_tus_upload_credentials,
+)
+from lib.api.mcp.tools.workflows import run_workflow  # noqa: E402
 
 _mcp_auth_mod.create_mcp_auth = _orig
 
@@ -48,78 +56,78 @@ def _make_user(email: str = "alice@example.com", name: str = "Alice") -> MagicMo
     return user
 
 
-# --- _build_project_url ---
+# --- build_project_url ---
 
 
 def test_build_project_url():
-    url = _build_project_url("abc-123")
+    url = build_project_url("abc-123")
     assert url.endswith("/projects/abc-123")
     assert not url.endswith("//projects/abc-123")
 
 
 def test_build_project_url_strips_trailing_slash():
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.FRONTEND_URL = "https://app.example.com/"
-        result = _build_project_url("proj-1")
+        result = build_project_url("proj-1")
     assert result == "https://app.example.com/projects/proj-1"
 
 
-# --- _build_settings_url ---
+# --- build_settings_url ---
 
 
 def test_build_settings_url():
-    url = _build_settings_url()
+    url = build_settings_url()
     assert url.endswith("/account")
     assert not url.endswith("//account")
 
 
 def test_build_settings_url_strips_trailing_slash():
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.FRONTEND_URL = "https://app.example.com/"
-        result = _build_settings_url()
+        result = build_settings_url()
     assert result == "https://app.example.com/account"
 
 
-# --- _require_api_key ---
+# --- require_api_key ---
 
 
 def test_require_api_key_passes_when_user_has_encrypted_key():
     user = _make_user()
     user.encrypted_openai_api_key = "some-encrypted-value"
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.OPENAI_API_KEY = None
-        _require_api_key(user)  # should not raise
+        require_api_key(user)  # should not raise
 
 
 def test_require_api_key_passes_when_env_has_key():
     user = _make_user()
     user.encrypted_openai_api_key = None
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.OPENAI_API_KEY = "sk-env-key"
-        _require_api_key(user)  # should not raise
+        require_api_key(user)  # should not raise
 
 
 def test_require_api_key_raises_when_no_key_anywhere():
     user = _make_user()
     user.encrypted_openai_api_key = None
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.OPENAI_API_KEY = None
         mock_cfg.FRONTEND_URL = "https://app.example.com"
         with pytest.raises(ValueError, match="Settings page"):
-            _require_api_key(user)
+            require_api_key(user)
 
 
 def test_require_api_key_error_message_contains_email():
     user = _make_user(email="carol@example.com")
     user.encrypted_openai_api_key = None
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.OPENAI_API_KEY = None
         mock_cfg.FRONTEND_URL = "https://app.example.com"
         with pytest.raises(ValueError, match="carol@example.com"):
-            _require_api_key(user)
+            require_api_key(user)
 
 
-# --- _resolve_user ---
+# --- resolve_user ---
 
 
 @pytest.mark.asyncio
@@ -128,10 +136,10 @@ async def test_resolve_user_calls_get_or_create():
     token = _make_token()
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        result = await _resolve_user(token)
+        result = await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="alice@example.com", name="Alice")
     assert result is mock_user
@@ -143,7 +151,7 @@ async def test_resolve_user_raises_when_email_missing():
     token.claims = {"name": "No Email"}
 
     with pytest.raises(RuntimeError, match="Token missing 'email' claim"):
-        await _resolve_user(token)
+        await resolve_user(token)
 
 
 @pytest.mark.asyncio
@@ -153,10 +161,10 @@ async def test_resolve_user_falls_back_email_as_name():
     mock_user = _make_user(email="bob@example.com", name="bob@example.com")
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        await _resolve_user(token)
+        await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="bob@example.com", name="bob@example.com")
 
@@ -169,10 +177,10 @@ async def test_resolve_user_azure_preferred_username_top_level():
     mock_user = _make_user(email="jane@example.com", name="Jane Doe")
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        result = await _resolve_user(token)
+        result = await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="jane@example.com", name="Jane Doe")
     assert result is mock_user
@@ -186,10 +194,10 @@ async def test_resolve_user_azure_preferred_username_no_name():
     mock_user = _make_user(email="jane@example.com", name="jane@example.com")
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        result = await _resolve_user(token)
+        result = await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="jane@example.com", name="jane@example.com")
     assert result is mock_user
@@ -203,10 +211,10 @@ async def test_resolve_user_azure_upstream_claims_email():
     mock_user = _make_user(email="jane@example.com", name="Jane Doe")
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        result = await _resolve_user(token)
+        result = await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="jane@example.com", name="Jane Doe")
     assert result is mock_user
@@ -220,10 +228,10 @@ async def test_resolve_user_azure_upstream_claims_preferred_username():
     mock_user = _make_user(email="jane@example.com", name="jane@example.com")
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        result = await _resolve_user(token)
+        result = await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="jane@example.com", name="jane@example.com")
     assert result is mock_user
@@ -241,10 +249,10 @@ async def test_resolve_user_email_takes_priority_over_preferred_username():
     mock_user = _make_user(email="official@example.com", name="Jane Doe")
 
     with patch(
-        "lib.api.mcp.get_or_create_user_by_email",
+        "lib.api.mcp.helpers.get_or_create_user_by_email",
         new=AsyncMock(return_value=mock_user),
     ) as mock_get:
-        result = await _resolve_user(token)
+        result = await resolve_user(token)
 
     mock_get.assert_awaited_once_with(email="official@example.com", name="Jane Doe")
     assert result is mock_user
@@ -256,7 +264,7 @@ async def test_resolve_user_raises_when_no_email_anywhere():
     token.claims = {"name": "No Email User", "upstream_claims": {"name": "No Email"}}
 
     with pytest.raises(RuntimeError, match="Token missing 'email' claim"):
-        await _resolve_user(token)
+        await resolve_user(token)
 
 
 @pytest.mark.asyncio
@@ -265,7 +273,7 @@ async def test_resolve_user_raises_when_preferred_username_not_email():
     token.claims = {"preferred_username": "jdoe"}
 
     with pytest.raises(RuntimeError, match="not a valid email address"):
-        await _resolve_user(token)
+        await resolve_user(token)
 
 
 # --- create_project ---
@@ -281,13 +289,13 @@ async def test_create_project_returns_ids_and_url():
     ctx = AsyncMock()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.create_project_record",
+            "lib.api.mcp.tools.projects.create_project_record",
             new=AsyncMock(return_value=project),
         ),
         patch(
-            "lib.api.mcp.finalize_file",
+            "lib.api.mcp.tools.projects.finalize_file",
             new=AsyncMock(return_value=file_record),
         ),
     ):
@@ -308,9 +316,9 @@ async def test_create_project_returns_project_url():
     project.id = uuid4()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.create_project_record", new=AsyncMock(return_value=project)),
-        patch("lib.api.mcp.finalize_file", new=AsyncMock(return_value=MagicMock(id=uuid4()))),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.projects.create_project_record", new=AsyncMock(return_value=project)),
+        patch("lib.api.mcp.tools.projects.finalize_file", new=AsyncMock(return_value=MagicMock(id=uuid4()))),
     ):
         result = await create_project(
             title="T", content_markdown="md", ctx=AsyncMock(), token=_make_token()
@@ -326,7 +334,7 @@ async def test_create_project_returns_project_url():
 @pytest.mark.asyncio
 async def test_run_workflow_rejects_invalid_type():
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=_make_user())),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=_make_user())),
     ):
         with pytest.raises(ValueError, match="Unknown workflow_type 'nonexistent'"):
             await run_workflow(
@@ -342,12 +350,12 @@ async def test_run_workflow_delegates_to_blocking_runner():
     project_json = json.dumps({"id": "p1", "project_url": "http://x/projects/p1"})
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.run_multiple_workflows_blocking", new=AsyncMock()
+            "lib.api.mcp.tools.workflows.run_multiple_workflows_blocking", new=AsyncMock()
         ) as mock_run,
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(return_value=project_json),
         ),
     ):
@@ -367,12 +375,12 @@ async def test_run_workflow_passes_approve_human_steps_to_runner():
     project_json = json.dumps({"id": "p1", "project_url": "http://x/projects/p1"})
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.run_multiple_workflows_blocking", new=AsyncMock()
+            "lib.api.mcp.tools.workflows.run_multiple_workflows_blocking", new=AsyncMock()
         ) as mock_run,
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(return_value=project_json),
         ),
     ):
@@ -400,13 +408,13 @@ async def test_run_workflow_returns_human_approval_required_payload():
     )
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.run_multiple_workflows_blocking",
+            "lib.api.mcp.tools.workflows.run_multiple_workflows_blocking",
             new=AsyncMock(side_effect=err),
         ),
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(),
         ) as mock_details,
     ):
@@ -432,12 +440,12 @@ async def test_run_workflow_passes_approve_web_search_to_runner():
     project_json = json.dumps({"id": "p1", "project_url": "http://x/projects/p1"})
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.run_multiple_workflows_blocking", new=AsyncMock()
+            "lib.api.mcp.tools.workflows.run_multiple_workflows_blocking", new=AsyncMock()
         ) as mock_run,
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(return_value=project_json),
         ),
     ):
@@ -465,13 +473,13 @@ async def test_run_workflow_returns_web_search_required_payload():
     )
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.run_multiple_workflows_blocking",
+            "lib.api.mcp.tools.workflows.run_multiple_workflows_blocking",
             new=AsyncMock(side_effect=err),
         ),
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(),
         ) as mock_details,
     ):
@@ -496,8 +504,8 @@ async def test_run_workflow_raises_when_no_api_key():
     user.encrypted_openai_api_key = None
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.env_config") as mock_cfg,
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.env_config") as mock_cfg,
     ):
         mock_cfg.OPENAI_API_KEY = None
         mock_cfg.FRONTEND_URL = "https://app.example.com"
@@ -520,9 +528,9 @@ async def test_list_projects_returns_project_list():
     project_item.project.title = "My Project"
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp.get_user_projects",
+            "lib.api.mcp.tools.projects.get_user_projects",
             new=AsyncMock(return_value=[project_item]),
         ),
     ):
@@ -540,8 +548,8 @@ async def test_list_projects_returns_empty_list_when_no_projects():
     user = _make_user()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_user_projects", new=AsyncMock(return_value=[])),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.projects.get_user_projects", new=AsyncMock(return_value=[])),
     ):
         raw = await list_projects(token=_make_token())
 
@@ -570,13 +578,13 @@ async def test_export_project_docx_returns_file():
     fake_docx = b"PK\x03\x04fake-docx-bytes"
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(MagicMock(), MagicMock()))),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(MagicMock(), MagicMock()))),
         patch(
-            "lib.api.mcp.generate_docx",
+            "lib.api.mcp.tools.files.generate_docx",
             new=AsyncMock(return_value=("/tmp/report.docx", "report_comments.docx")),
         ),
-        patch("lib.api.mcp.aiofiles.open", _make_aiofiles_open_mock(fake_docx)),
+        patch("lib.api.mcp.tools.files.aiofiles.open", _make_aiofiles_open_mock(fake_docx)),
     ):
         result = await export_project_docx(project_id=project_id, token=_make_token())
 
@@ -593,13 +601,13 @@ async def test_export_project_docx_passes_filters_to_service():
     project_id = str(uuid4())
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(MagicMock(), MagicMock()))),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(MagicMock(), MagicMock()))),
         patch(
-            "lib.api.mcp.generate_docx",
+            "lib.api.mcp.tools.files.generate_docx",
             new=AsyncMock(return_value=("/tmp/out.docx", "out_comments.docx")),
         ) as mock_gen,
-        patch("lib.api.mcp.aiofiles.open", _make_aiofiles_open_mock(b"")),
+        patch("lib.api.mcp.tools.files.aiofiles.open", _make_aiofiles_open_mock(b"")),
     ):
         await export_project_docx(
             project_id=project_id,
@@ -622,9 +630,9 @@ async def test_get_project_returns_details():
     expected = json.dumps({"id": "p1", "title": "Test"})
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(return_value=expected),
         ) as mock_details,
     ):
@@ -662,10 +670,10 @@ async def test_list_project_files_returns_files_with_reference():
     match = ReferenceFileMatch(reference_id=ref_id, file_id=str(file1.id), source=MatchSource.MANUAL_UPLOAD)
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
-        patch("lib.api.mcp.get_project_files_list_items", new=AsyncMock(return_value=[file1])),
-        patch("lib.api.mcp.get_file_reference_matches", new=AsyncMock(return_value=[match])),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.tools.files.get_project_files_list_items", new=AsyncMock(return_value=[file1])),
+        patch("lib.api.mcp.tools.files.get_file_reference_matches", new=AsyncMock(return_value=[match])),
     ):
         raw = await list_project_files(project_id=project_id, token=_make_token())
 
@@ -695,10 +703,10 @@ async def test_list_project_files_null_reference_when_unmatched():
     file1.revision = None
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
-        patch("lib.api.mcp.get_project_files_list_items", new=AsyncMock(return_value=[file1])),
-        patch("lib.api.mcp.get_file_reference_matches", new=AsyncMock(return_value=[])),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.tools.files.get_project_files_list_items", new=AsyncMock(return_value=[file1])),
+        patch("lib.api.mcp.tools.files.get_file_reference_matches", new=AsyncMock(return_value=[])),
     ):
         raw = await list_project_files(project_id=str(project.id), token=_make_token())
 
@@ -714,10 +722,10 @@ async def test_list_project_files_returns_empty_list():
     project.current_revision = 1
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
-        patch("lib.api.mcp.get_project_files_list_items", new=AsyncMock(return_value=[])),
-        patch("lib.api.mcp.get_file_reference_matches", new=AsyncMock(return_value=[])),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.tools.files.get_project_files_list_items", new=AsyncMock(return_value=[])),
+        patch("lib.api.mcp.tools.files.get_file_reference_matches", new=AsyncMock(return_value=[])),
     ):
         raw = await list_project_files(project_id=str(project.id), token=_make_token())
 
@@ -736,10 +744,10 @@ async def test_remove_reference_file_returns_file_and_unlinked_refs():
     ref_id = str(uuid4())
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
         patch(
-            "lib.api.mcp.delete_project_file_with_cleanup",
+            "lib.api.mcp.tools.files.delete_project_file_with_cleanup",
             new=AsyncMock(return_value=(1, [ref_id])),
         ),
     ):
@@ -760,10 +768,10 @@ async def test_remove_reference_file_raises_when_not_found():
     file_id = str(uuid4())
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
         patch(
-            "lib.api.mcp.delete_project_file_with_cleanup",
+            "lib.api.mcp.tools.files.delete_project_file_with_cleanup",
             new=AsyncMock(return_value=(0, [])),
         ),
     ):
@@ -781,10 +789,10 @@ async def test_remove_reference_file_with_no_associations():
     file_id = str(uuid4())
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.files.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
         patch(
-            "lib.api.mcp.delete_project_file_with_cleanup",
+            "lib.api.mcp.tools.files.delete_project_file_with_cleanup",
             new=AsyncMock(return_value=(1, [])),
         ),
     ):
@@ -796,20 +804,20 @@ async def test_remove_reference_file_with_no_associations():
     assert data["removed_from_reference_ids"] == []
 
 
-# --- _build_tus_url ---
+# --- build_tus_url ---
 
 
 def test_build_tus_url_strips_mcp_suffix():
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.MCP_BASE_URL = "https://app.example.com/mcp"
-        result = _build_tus_url()
+        result = build_tus_url()
     assert result == "https://app.example.com/tus"
 
 
 def test_build_tus_url_handles_trailing_slash_after_stripping():
-    with patch("lib.api.mcp.env_config") as mock_cfg:
+    with patch("lib.api.mcp.helpers.env_config") as mock_cfg:
         mock_cfg.MCP_BASE_URL = "https://app.example.com/mcp"
-        result = _build_tus_url()
+        result = build_tus_url()
     assert not result.endswith("//tus")
     assert result.endswith("/tus")
 
@@ -827,9 +835,9 @@ async def test_get_tus_upload_credentials_returns_required_fields():
     project_id = str(uuid4())
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
-        patch("lib.api.mcp.env_config") as mock_cfg,
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.uploads.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.helpers.env_config") as mock_cfg,
     ):
         mock_cfg.MCP_BASE_URL = "https://app.example.com/mcp"
         mock_cfg.AUTH_SECRET = "test-secret"
@@ -862,9 +870,9 @@ async def test_get_tus_upload_credentials_includes_reference_id():
     ref_id = str(uuid4())
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
-        patch("lib.api.mcp.env_config") as mock_cfg,
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.uploads.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.helpers.env_config") as mock_cfg,
     ):
         mock_cfg.MCP_BASE_URL = "https://app.example.com/mcp"
         mock_cfg.AUTH_SECRET = "test-secret"
@@ -885,9 +893,9 @@ async def test_get_tus_upload_credentials_bearer_token_expires_in_one_hour():
     user = _make_user()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(MagicMock(), MagicMock()))),
-        patch("lib.api.mcp.env_config") as mock_cfg,
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.uploads.get_project_access", new=AsyncMock(return_value=(MagicMock(), MagicMock()))),
+        patch("lib.api.mcp.helpers.env_config") as mock_cfg,
     ):
         mock_cfg.MCP_BASE_URL = "https://app.example.com/mcp"
         mock_cfg.AUTH_SECRET = "test-secret"
@@ -921,9 +929,9 @@ async def test_create_project_without_content_returns_next_step():
     project.id = uuid4()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.create_project_record", new=AsyncMock(return_value=project)),
-        patch("lib.api.mcp.finalize_file", new=AsyncMock()) as mock_finalize,
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.projects.create_project_record", new=AsyncMock(return_value=project)),
+        patch("lib.api.mcp.tools.projects.finalize_file", new=AsyncMock()) as mock_finalize,
     ):
         raw = await create_project(title="Test", ctx=AsyncMock(), token=_make_token())
 
@@ -945,9 +953,9 @@ async def test_create_revision_with_content_returns_file_id():
     file_record.id = uuid4()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.create_new_revision", new=AsyncMock(return_value=(2, ["claim_extraction"]))),
-        patch("lib.api.mcp.finalize_file", new=AsyncMock(return_value=file_record)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.revisions.create_new_revision", new=AsyncMock(return_value=(2, ["claim_extraction"]))),
+        patch("lib.api.mcp.tools.revisions.finalize_file", new=AsyncMock(return_value=file_record)),
     ):
         raw = await create_revision(
             project_id=str(uuid4()),
@@ -967,9 +975,9 @@ async def test_create_revision_without_content_returns_next_step():
     user = _make_user()
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.create_new_revision", new=AsyncMock(return_value=(3, []))),
-        patch("lib.api.mcp.finalize_file", new=AsyncMock()) as mock_finalize,
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.revisions.create_new_revision", new=AsyncMock(return_value=(3, []))),
+        patch("lib.api.mcp.tools.revisions.finalize_file", new=AsyncMock()) as mock_finalize,
     ):
         raw = await create_revision(
             project_id=str(uuid4()),
@@ -1006,9 +1014,9 @@ async def test_list_revisions_returns_all_revisions():
     file_rev2.created_at = MagicMock(isoformat=MagicMock(return_value="2026-04-10T00:00:00"))
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
-        patch("lib.api.mcp.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
-        patch("lib.api.mcp.get_files_by_project_id", new=AsyncMock(return_value=[file_rev1, file_rev2])),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.tools.revisions.get_project_access", new=AsyncMock(return_value=(project, MagicMock()))),
+        patch("lib.api.mcp.tools.revisions.get_files_by_project_id", new=AsyncMock(return_value=[file_rev1, file_rev2])),
     ):
         raw = await list_revisions(project_id=str(project.id), token=_make_token())
 
@@ -1031,9 +1039,9 @@ async def test_get_project_with_revision_passes_revision():
     expected = json.dumps({"id": "p1", "title": "Test", "revision": 1})
 
     with (
-        patch("lib.api.mcp._resolve_user", new=AsyncMock(return_value=user)),
+        patch("lib.api.mcp.helpers.resolve_user", new=AsyncMock(return_value=user)),
         patch(
-            "lib.api.mcp._get_project_details_json",
+            "lib.api.mcp.serialization.get_project_details_json",
             new=AsyncMock(return_value=expected),
         ) as mock_details,
     ):
